[agents] Move all ACP-related code to agents-features-acp
