--- agents/agents-acp/Module.md
@@ -1,1 +1,0 @@
-# Module agents-acp
