+++ koog-agents/build.gradle.kts
@@ -47,7 +47,6 @@ val included = setOf(
     ":agents:agents-features:agents-features-trace",
     ":agents:agents-features:agents-features-tokenizer",
     ":agents:agents-features:agents-features-snapshot",
-    ":agents:agents-acp",
     ":agents:agents-mcp",
     ":agents:agents-tools",
     ":agents:agents-utils",
